nextjs15-fastapi/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.3",
    "python-dotenv>=1.0.1",
    "supabase>=2.9.1",
    "uvicorn>=0.32.0",
]

[tool.uv]
dev-dependencies = [
    "mypy>=1.13.0",
]
